Moved generateXMLAssets() methods to MyXMLReader
